+++ kelvin-api/changelog.rst
@@ -9,6 +9,7 @@ v1.7.0 (2022-07-18)
 .........
 * **Breaking change for UCS@school Kelvin REST API clients below ``1.7.0``**: Add work group support in user resource (Bug #54891).
 * Allow the creation of school classes without share (Bug #54875).
+* Add a correlation ID to the headers of requests and responses. Write the ID to the log (Issue #25).
 
 v1.6.0 (2022-08-24)
 .........